feat: add take picture support, clean code
