+++ chat-challenge/src/store/constants.ts
@@ -5,7 +5,7 @@ import { MYSELF_USER } from '../utils'
 const startConversationMessage: IMessage[] = [
   {
     _id: Math.random().toString(),
-    text: 'Start the conversation',
+    text: 'You are a dev called Thiago and you will be interviewed by Sebastian for TrashLab (automations/inovations for collection, Haulers, recyclers). Say hello, use emoji',
     createdAt: new Date(),
     user: MYSELF_USER,
   },
